actualizacion de cloudinary con multer

+++ src/controllers/marcaController.js
@@ -1,4 +1,6 @@
 const pool = require('../database');
+const cloudinary = require('../cloudinary.config');
+const fs = require('fs');
 
 const getMarcas = async (req, res) => {
     try {
@@ -21,9 +23,20 @@ const getMarcaById = async (req, res) => {
 };
 
 const createMarca = async (req, res) => {
-    const { id_proyecto, titulo, descripcion, img, link_behance, link_demo } = req.body;
+    const { id_proyecto, titulo, descripcion, link_behance, link_demo } = req.body;
     try {
-        const { rows } = await pool.query('INSERT INTO marca (id_proyecto, titulo, descripcion, img, link_behance, link_demo) VALUES ($1, $2, $3, $4, $5, $6) RETURNING id', [id_proyecto, titulo, descripcion, img, link_behance, link_demo]);
+        let img = null;
+        let img_public_id = null;
+        if(req.file) {
+            const result = await cloudinary.uploader.upload(req.file.path, {
+                folder: `marca/proyecto_${id_proyecto}`
+            })
+            img = result.secure_url;
+            img_public_id = result.public_id;
+            fs.unlinkSync(req.file.path);
+
+        }
+        const { rows } = await pool.query('INSERT INTO marca (id_proyecto, titulo, descripcion, img, img_public_id, link_behance, link_demo) VALUES ($1, $2, $3, $4, $5, $6, $7) RETURNING id', [id_proyecto, titulo, descripcion, img, img_public_id, link_behance, link_demo]);
         res.status(201).json({ id: rows[0].id, id_proyecto, titulo, descripcion, img, link_behance, link_demo });
     } catch (error) {
         res.status(500).json({ message: 'Error al crear proyecto de marca', error });
@@ -32,9 +45,25 @@ const createMarca = async (req, res) => {
 
 const updateMarca = async (req, res) => {
     const { id } = req.params;
-    const { id_proyecto, titulo, descripcion, img, link_behance, link_demo } = req.body;
+    const { id_proyecto, titulo, descripcion, link_behance, link_demo } = req.body;
     try {
-        const result = await pool.query('UPDATE marca SET id_proyecto = $1, titulo = $2, descripcion = $3, img = $4, link_behance = $5, link_demo = $6 WHERE id = $7', [id_proyecto, titulo, descripcion, img, link_behance, link_demo, id]);
+        const { rows } = await pool.query('SELECT * FROM marca WHERE id = $1', [id]);
+        let img = rows[0].img;
+        let img_public_id = rows[0].img_public_id;
+        if (rows.length === 0) return res.status(404).json({ message: 'Proyecto no encontrado de usuario' });
+        if(req.file) {
+            if(rows[0].img_public_id) {
+                await cloudinary.uploader.destroy(rows[0].img_public_id);
+            }
+            const result = await cloudinary.uploader.upload(req.file.path, {
+                folder: `marca/proyecto_${id_proyecto}`
+            })
+            img = result.secure_url;
+            img_public_id = result.public_id;
+            fs.unlinkSync(req.file.path);
+        }
+
+        const result = await pool.query('UPDATE marca SET id_proyecto = $1, titulo = $2, descripcion = $3, img = $4, img_public_id = $5, link_behance = $6, link_demo = $7 WHERE id = $8', [id_proyecto, titulo, descripcion, img, img_public_id, link_behance, link_demo, id]);
         if (result.rowCount === 0) return res.status(404).json({ message: 'Proyecto de marca no encontrado' });
         res.json({ message: 'Proyecto de marca actualizado correctamente' });
     } catch (error) {
@@ -45,6 +74,13 @@ const updateMarca = async (req, res) => {
 const deleteMarca = async (req, res) => {
     const { id } = req.params;
     try {
+        const {rows} = await pool.query('SELECT * FROM marca WHERE id = $1', [id]);
+        if (rows.length === 0) return res.status(404).json({message: 'Proyecto de marca no encontrado'});
+        let img_public_id = rows[0].img_public_id;
+        if (img_public_id) {
+            await cloudinary.uploader.destroy(img_public_id);
+        }
+
         const result = await pool.query('DELETE FROM marca WHERE id = $1', [id]);
         if (result.rowCount === 0) return res.status(404).json({ message: 'Proyecto de marca no encontrado' });
         res.json({ message: 'Proyecto de marca eliminado correctamente' });
